+++ src/Exceptions/NotFoundException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Netflex\Query\Exceptions;
+
+class NotFoundException extends QueryBuilderException
+{
+  public function __construct()
+  {
+    parent::__construct('Not found');
+  }
+}
